Switch from Jest to Vitest
At the time of this commit, Jest doesn't support loading {wasm} files
natively and there is no simple workaround for it. We need to write a
custom sync transformer for Jest to handle {wasm} files correctly, but
the Node {WebAssembly} apis are async and so we cannot write such a
transformer easily.

Switch to Vitest which supports loading {wasm} files. It doesn't support
importing {wasm} files as ESM modules natively, but using a Vite plugin
called {vite-plugin-wasm} we are able to do so.

The {ergo-lib-wasm-browser} package is added to inline deps of Vitest
config, so that Vitest apply transformations on it (which causes the
{wasm} files to be imported correctly). By default, Vitest doesn't apply
any transformations on {node_modules}.

In addition to {vite-plugin-wasm}, we also need another plugin called
{vite-plugin-top-level-await}, which is complementary to the former plugin.

The environment for Vitest is set to {happy-dom} so that there is no need
to mock some of browser-specific apis which will be needed in the future.

Also update the {test} npm script to run {vitest} instead of {react-scripts}.

In addition, remove all jest related configurations and files, including:
- {react-app/jest} eslint configuration
- All jest related packages
- {setupTests} file