Add a patch to fix {fetch} issue in {vite-plugin-wasm}
When the tests are run, we get an undefined {fetch} error. That's because
the plugin uses {fetch} api inside, which isn't supported in Node. Even
if we mock the {fetch} api using some libraries like {node-fetch} or
{cross-fetch}, the apis are still incompatible and the error still exists
(in some other form).

To fix the error, the implementation of the plugin should change so that
if we are running in Node, it uses native {fs} apis instead.

To do so, {patch-package} tool is used. Also add a {postinstall} npm script
to apply the patch after installing dependencies.

Link to related issue in {vite-plugin-wasm} repo:
https://github.com/Menci/vite-plugin-wasm/issues/4