Fix Vitest issue with Typeorm
By default, Vitest externalize all packages in {node_modules} to use Vite
SSR transform pipeline, which is intended to be used in Node. Current
version of {typeorm}, on the other hand, has some critical issues with
being run in Node directly (because of some bugs in {exports} field of
its {package.json}). To opt out {typeorm} being loaded by Node directly,
we need to first add {typeorm} to Vitest inline deps (so that it passes
through vite transformers) and also change the transform mode of all js
and ts files to {web}, so that all of them are benefited from Vite default
browser-oriented transformations. As {typeorm} runs in browser as expected
and its {exports} field in its {package.json} are correct for browser,
the issue is solved.

Note that even if we upgrade {typeorm} to the current latest (which has
been solved the {exports} issue for Node), there still exists some other
issues which cannot be resolved without the solution specified. So applying
Vite browser transformations to {typeorm} seems inevitable (at least for
now).